craftstore/pyproject.toml
[tool.poetry]
name = "craftstoreapp"
version = "0.0.1"
description = "BETA"
authors = ["Robitnik"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
cryptography = "^43.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
